Update patch set 32
Patch Set 32:

(1 comment)

Set 32 has some very specific hacks in it for a quick test of the DPDK interface, it is not sufficient for the general read/write case - that will require reduceLib integration which will start this week.  This code supports proving the DPDK i/f by stepping through bdevio doing 1 4K write and examining src/dst in gdb followed by a 4K read (with mbufs set to the compressed length of the 4K write) and confirming original data in the dest buffer which currently will be in the io ctx of the completed read (not the read buffer).

Patch-set: 32
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Work-in-progress: false