Update patch set 1
Patch Set 1: -Code-Review

Yes, you are right. In the next patch you are suspending the IO queue anyway so the lun->tasks queue could be used not only for tasks that waiting for completion but also not started yet because reset task is in queue. So instead of placing it in top of the queue just insert tail and use it as barier.

The when this reset task is finished it can trigger execution of next tasks in the queue.

But of course I can wrong here. Removing my -1 to not confuse ether reviewers.

Patch-set: 1
Reviewer: Pawel Wodkowski <1011230@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0