vhost: enable spdk recovering from crash or killing
When guest system is running, if spdk crashes or killed
by signals, any access to the block dev will hung. So
when spdk recovers, it needs to resubmit io requests and
let the block dev recover.

To support this, the spdk should maintain an array of the
inflight io requests status to recover the unfinished io.
The arrays lays in the mmap memory between qemu and spdk,
so when spdk crashes, the memory still exists. At the
begining of the initiation of spdk, it will resubmit unfinished
requests at first, and then go on with the normal poller work.

Change-Id: I0f5433e846bcd34a00873085d1b16c7892dfcba7
Signed-off-by: lilin24 <lilin24@baidu.com>