vhost: enable spdk recovering from crash or killing
When guest system is running, if spdk crashes or killed
by signals, access to the block dev may hung.

Because IO processing is out of order, We can't determine
which IOs are actually completed and which IOs are in
inflight. In other words, the order of IO recorded in avail
and used arrays may be different.This may lead to the
resubmission of some completed IOs, while other inflight IOs
are not submitted, resulting in IO hung.

To support this, the spdk should maintain an array of the
inflight io requests status to recover the unfinished io.
The arrays lays in the mmap memory between qemu and spdk,
so when spdk crashes, the memory still exists. At the
begining of the initiation of spdk, it will resubmit unfinished
requests at first, and then go on with the normal poller work.

Change-Id: I0f5433e846bcd34a00873085d1b16c7892dfcba7
Signed-off-by: lilin24 <lilin24@baidu.com>
Signed-off-by: Ni Xun <nixun@baidu.com>
Signed-off-by: Zhang Yu <zhangyu31@baidu.com>