Update patch set 14
Patch Set 14:

Hi Jim, In my mind, it should be put in transport ID (since it is optional). If we put that, it will be transparent for any other applications, like nvme fio plugin or bdev fio plugin. It will not need those applications to provide additional command line. I think that put in the controller opts are also not convincing. The digest is only related with the TCP transport...

Patch-set: 14
Reviewer: Ziye Yang <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>