Update patch set 1
Patch Set 1: Code-Review-1

Should we remove the extra Reviewed-by, etc. from the commit message when it was merged to master?

I don't think we've always been consistent on this, but would like to moving forward.

I think you keep Reviewed-on, but remove everything after it.  Modify the Reviewed-on to add "(master)" after the URL.  When this patch gets merged, another Reviewed-on will be added, this time for the 18.10.1 patch.

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1