bdev/compress: add reduce_vol_load
WIP, moving this confile file support ripping to the initial compress
bvdev patch

Also removes all config file support that was in there as PT module
was used as a starting point.  There is no config file support
for compress, RPCs for create and delete and on disk metadata is
used for examine().

Change-Id: Ib9cd7202395980ec59e0d6b37aca2c5a20a76a52
Signed-off-by: paul luse <paul.e.luse@intel.com>