Update patch set 3
Patch Set 3:

> stupid question as I'm not really sure what I'm looking at :) . Who
 > is paying attention to the new param?

The nvmftestinit function takes it in. If it is equal to iso, it does all of the initial prep stuff that autotest normally takes care of. It's really useful when trying to debug an individual script from your desk.

Patch-set: 3
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>