nvmf/tcp: fix the recv state setting if no shared buffer available
Fix the issue when there is no data buffer allocated,
the previous method is wrong to set the recv pdu state.

The reason is that: When there is no data buffer allocated,
we still need to handle the incoming pdu. And when there is
buffer, we should resume the req handling with buffer, but
we should not set the pdu receving state.

Change-Id: I1cc2723acc7b0a17407c3a2e6273313a4e612916
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>