Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
HiLiu changpeng
  Because IO processing is out of order, if we do not record the inflight status of IO, We can't determine which IOs are actually completed 
and which IOs are in inflight. In other words, the order of IO recorded in avail and used arrays may be different.
If we only use last_avail_idx and last_used_id to get IO that needs to be resubmitted, this may lead to the resubmission of some completed IOs, 
while other inflight IOs are not submitted, resulting in IO hung.

Patch-set: 2
Reviewer: Lin Li <1015052@d5d70762-12d0-45a1-890d-524b12d3f735>