Update patch set 11
Patch Set 11:

The 5% percent variance is for above and below, so that we have a range of 0.95 to 1.05. The base is same from MiB counted value. The actual data bytes transfered is also counted in MiB. May refer to 
https://ci.spdk.io/spdk/builds/review/ba223b01e30406f0c4678fa862de4d3285c81a04.1543206563/fedora-02/build.log

 23:36:04	# end_io_count=244637696
 23:36:04	# read_result=10465280
 23:36:04	# '[' true = true ']'
  23:36:04	# echo '10485760*0.95 < 10465280 && 10465280 < 10485760*1.05'

If up 5% is too much considering the major rate limiting work here, we can have a small one like 2%. While the down 5% could be necessary as this is to limit and restrict the IOs (and bandwidth).

These are my thoughts.

Patch-set: 11
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>