rpc/notifications: add notification support to RPC client and SPDK host
Client sends some number of notification requests to feed the SPDK
server with ID numbers it can response to. Responses will be returned
asynchronously when event occurs. In the meantime client can issue other
RPC requests to interact with SPDK.

Note that altough the implementation in C library is complete we are not
ready yet to do fully asynchronouse RPC conversation using our Python
client. But this is not stopping user from issuing two connections and
listen for notificatons on one of them.

Change-Id: Ia5288d83cc5e9f18ef1622d3f15b9fe3bbf640ed
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>