Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Hey Lance,

You're right that the "master" parts of this need to change.  For this specific commit, we shouldn't use 18.10.1 yet though - just 18.10.1-pre (and v18.10 to point to the head of the v18.10.x branch).

When we do the patch for the 18.10.1 tag itself, this file would be updated to change the version to 18.10.1, and the Source URL to the v18.10.1 tag.

-Jim

Patch-set: 3
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1