Update patch set 5
Patch Set 5:

In reality, all configuration operations happen on the SPDK master core, so would be impossible to hit an issue with main SPDK use cases.  But these libraries are designed to be used outside of main SPDK framework, so this patch is a good one to add.

Patch-set: 5
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>