test: add mem_callbacks unit test
This will be helpful during development of upstream
DPDK patches to ensure that buffers do not cross
memory callback allocation boundaries, and that memory
freed by DPDK is in the same units as it was allocated.

Since DPDK currently will fail these tests, do not
plumb them into unittest.sh - but still build it by
default.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I5b68e9f0864304962b9189b9f5e58f828f0d0e54