util/string: sprintf_realloc to concatenate strings with realloc
Appending string by using sprintf with realloc will be generally
usable and add sprintf_realloc() and vsprintf_realloc() to the utility.

The behavior of sprintf_realloc() and vsprintf_realloc() are slightly
different from libc's realloc(). The input buffer will be freed in any
case and should not be used or passed to free().

Change-Id: I8b69d9640e86e1862ddd3917995bad6f59426b7e
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Signed-off-by: Chunyang Hui <chunyang.hui@intel.com>
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>