util/string: sprintf_realloc to concatenate strings with realloc
Appending string by using sprintf with realloc will be generally
usable and add sprintf_realloc() and vsprintf_realloc() to the utility.

The behavior of sprintf_realloc() and vsprintf_realloc() when they
fail is different from realloc(). When they fail, the original buffer
is freed by them and the caller for them must not call free() to
the original buffer when they fail. This difference simplify the
implementation of the user for sprintf_realloc() or vsprintf_realloc().

Change-Id: I8b69d9640e86e1862ddd3917995bad6f59426b7e
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Signed-off-by: Chunyang Hui <chunyang.hui@intel.com>
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>