rdma: after rdma update ibv state we check the state value
We found that state value may be unreasonable, so befor we
use the state value we do some judgement. The unreasonable state
probably means hardware issue, so the process flow become unpredicatable.
assert() may be the best choice. Also in group add path, we need to remove
TATILQ when we want to destroy qpair.

Change-Id: I213f4d684b103cce7bc072aecd591e2c491e0596
Signed-off-by: JinYu <jin.yu@intel.com>