jsonrpc: fix received JSON buffer overwrite
The receive buffer is shared across all requests on one connection. If
RPC handlers do the lazy decode then content of the buffer might be
overwritten by now. As we don't have any requirement here we must assert
that the receive buffer is valid till response is produced.

Fix this issue by copying request data and work on the copy.

Change-Id: I665be446cbcd8f625e5a73514582efad3021a4ff
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>