nvmf: fix the "queued" list initialized bug.
The error situation is that:

When we use rpc to create/delete the subsystem.
we call spdk_nvmf_poll_group_add_subsystem/
spdk_nvmf_poll_group_add_subsystem.

If with the same sid, I think that the "queued"
list will be re-intiliazed again.

So I would like to propose this patch to enable
the initializaiton of this list early

Change-Id: Ic3e590163c91cdb9b8c3447ac85eae7ba4f876b7
Signed-off-by: Ziye Yang <ziye.yang@intel.com>