nvmf: fix the queued not initialized bug.
The error situation is that:

When a subsystem is down and removed, and the host
wants to reconnect, that time, it checks the sgroup
state (it is in INACTIVE STATE), and the qpair could be
not initialzied.

So I think that we need to always initilize the queued list
early. This could prevent lots of issues.

Change-Id: Ic3e590163c91cdb9b8c3447ac85eae7ba4f876b7
Signed-off-by: Ziye Yang <ziye.yang@intel.com>