nvmf/rdma: fix the coredump issue when ctrlr + c target
When the host connects the target and does the fio,
if we use ctrlr + c, it will be crash. The issue
is that we found the rqpair->qpair.group is NULL,
when caling spdk_thread_send_msg since qpair.group
is NULL.

And this patch can solve this issue, it means that
we should already call the single spdk_nvmf_qpair_disconnect
function, which will eliminate lots of issues.

Change-Id: Id36cfac2be9abc707bf75a2e1ddb3f414610b6f1
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>