Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

Now that I think this through, this might be a bug on the initiator side instead. Even if the target has a local buffer available to use for sending/receiving data, if the initiator is not calling recv on all of the sockets while it shuts down, it could still get into a case where the OS/NIC buffers are so backed up with data from other sockets that the target can't send to it. I think we need to fix the initiator.

Patch-set: 6
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1