Update patch set 6
Patch Set 6:

Also for the incapsule data, we still can not use that. Considering the following case, if the queue depth is 4, but the incapsule size is 4KB,which means that the total size of incapsule size is 16KB. However, if the data buffer request is 64KB (> 16KB), we still do not have the enough buffer to hold the data.

Patch-set: 6