Update patch set 7
Patch Set 7:

The version check was added specifically to keep pkgdep.sh from failing when it executes on a system in the TP.  When its run by a developer on their own system clearly they will know how to deal with failures, missing packages or whatever.  So somewhere in the history of this simple patch, LOL, it was commented that a version check isn't the best because of what I just stated, it's really only relevant for the TP systems which are going to change over time.  So, instead just ignore and keep going.

I don't know now though, maybe this is all just a waste - we have a trap in there that this patch "gets around" because in this case it didn't seem to matter much that the SPDK_CLI dependencies weren't available.  Maybe instead of this change we should change the trap to not exit on error but just print a message and keep on going.  So this patch would simply try to install everything and but we'd remove the "exit 1" from the trap and print a more descriptive message about a package not being available since that's what we're really catching with it now anyways.

Other thoughts?

Patch-set: 7
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>