Update patch set 7
Patch Set 7: Code-Review+2

So I'm fine with this patch as-is.  Relaxing the exit 1 would be nice, but the challenge is that even with this print that you've added, it's probably going to get lost in all of the other spew from apt-get.

I think if we end up relaxing the exit 1, we concatenate these 'error' messages and print them out at the very end.  Then the user sees them after all of the apt-get, dnf or whatever spew.

Patch-set: 7
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2