Update patch set 3
Patch Set 3: Code-Review-1

I am pretty sure this works as is. It scares me though because it assumes that they are right next to each other in the structure. This could change on accident for some reason in the future then we are tracking down memory corruption.
If you think that this will be a big performance improvement, I think you should make a struct nvme_cmd_rsp or something like that and embed the two objects in it. then memset that whole structure. That way, there is a language construct keeping the two together.

Patch-set: 3
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1