Update patch set 3
Patch Set 3: Code-Review-1

I'm pretty sure the assembly on the before and the after are going to come out more or less the same, because the cmd is 64 bytes already. If you compile the before and after and objdump it and show there is a meaningful difference in the results, I can get on board with this (although I'd recommend defining a structure like Seth suggested).

Patch-set: 3
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1