nvmf: set the qpair state to active in new_qpair function.
It is not exactly correct to set the state in
spdk_nvmf_poll_group_poll function. The reason is that this
function is called by spdk_nvmf_poll_group_poll with event
function.

And see the logic in spdk_nvmf_poll_group_poll,
the state SPDK_NVMF_QPAIR_ACTIVATING is useless. so I think
that put this state in new_qpair function is useful,
we can use this state to track the qpair state, if the qpair
is in SPDK_NVMF_QPAIR_ACTIVATING state, which means that
the qpair is not added into the related group yet. And if
there is disconnect operation when the qpair is in
SPDK_NVMF_QPAIR_ACTIVATING state, we can use this state to
defer the disconnect of this qpair.

Change-Id: I88ad724c898ce27e157df709f669c0e7d288a761
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>