nvmf: set the qpair state to ACTIVATING state early.
It is not exactly correct to set the state in
spdk_nvmf_poll_group_poll function. The reason is that this
function is called by spdk_nvmf_poll_group_poll with event
function.

And see the logic in spdk_nvmf_poll_group_poll,
the state SPDK_NVMF_QPAIR_ACTIVATING is useless. so I think
that we should set the state to SPDK_NVMF_QPAIR_ACTIVATING after
calling new_qpair function (cb_fn).

The reason is that: In the spdk_nvmf_qpair_disconnect function
(current code), we check qpair state == SPDK_NVMF_QPAIR_UNINITIALIZED,
we just free. However, there is a timing to change qpair from
SPDK_NVMF_QPAIR_UNINITIALIZED to SPDK_NVMF_QPAIR_ACTIVATING
(since it is entering another call). So if the qpair state is
SPDK_NVMF_QPAIR_UNINITIALIZED, we cannot make sure it is still
in the original thread or in the poll group's thread. So
I think that the safeway is to change the return type of new_qpair
function, which will be much more safe to guarantee the logic
in spdk_nvmf_qpair_disconnect function.

Change-Id: I88ad724c898ce27e157df709f669c0e7d288a761
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>